Refactored several classes into respective packages
