+++ api/src/main/java/org/openmrs/module/patientgrid/PatientGridUtils.java
@@ -21,6 +21,11 @@ import org.openmrs.Obs;
 import org.openmrs.api.APIException;
 import org.openmrs.api.context.Context;
 import org.openmrs.module.patientgrid.PatientGridColumn.ColumnDatatype;
+import org.openmrs.module.patientgrid.converter.PatientGridAgeConverter;
+import org.openmrs.module.patientgrid.definition.AgeAtLatestEncounterPatientDataDefinition;
+import org.openmrs.module.patientgrid.definition.LocationPatientDataDefinition;
+import org.openmrs.module.patientgrid.definition.ObsForLatestEncounterPatientDataDefinition;
+import org.openmrs.module.patientgrid.definition.PersonUuidDataDefinition;
 import org.openmrs.module.reporting.common.AgeRange;
 import org.openmrs.module.reporting.common.TimeQualifier;
 import org.openmrs.module.reporting.data.converter.AgeRangeConverter;
