+++ api/src/main/java/org/openmrs/module/patientgrid/converter/PatientGridAgeConverter.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.patientgrid;
+package org.openmrs.module.patientgrid.converter;
 
 import org.openmrs.module.reporting.common.Age;
 import org.openmrs.module.reporting.data.converter.DataConverter;
