+++ api/src/main/java/org/openmrs/module/patientgrid/evaluator/ObsForLatestEncounterPatientDataEvaluator.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.patientgrid;
+package org.openmrs.module.patientgrid.evaluator;
 
 import static org.openmrs.module.patientgrid.PatientGridConstants.KEY_MOST_RECENT_ENCS;
 
@@ -9,6 +9,8 @@ import org.openmrs.Encounter;
 import org.openmrs.EncounterType;
 import org.openmrs.Obs;
 import org.openmrs.annotation.Handler;
+import org.openmrs.module.patientgrid.definition.ObsForLatestEncounterPatientDataDefinition;
+import org.openmrs.module.patientgrid.PatientGridUtils;
 import org.openmrs.module.reporting.data.patient.EvaluatedPatientData;
 import org.openmrs.module.reporting.data.patient.definition.PatientDataDefinition;
 import org.openmrs.module.reporting.data.patient.evaluator.PatientDataEvaluator;
