+++ api/src/test/java/org/openmrs/module/patientgrid/evaluator/PersonUuidDataEvaluatorTest.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.patientgrid;
+package org.openmrs.module.patientgrid.evaluator;
 
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
@@ -6,6 +6,7 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 import org.openmrs.Cohort;
 import org.openmrs.api.PersonService;
+import org.openmrs.module.patientgrid.definition.PersonUuidDataDefinition;
 import org.openmrs.module.reporting.data.person.EvaluatedPersonData;
 import org.openmrs.module.reporting.data.person.service.PersonDataService;
 import org.openmrs.module.reporting.evaluation.EvaluationContext;
